backend/requirements.txt
asgiref==3.9.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
defusedxml==0.7.1
Django==5.2.6
django-cors-headers==4.4.0
django-js-asset==3.1.2
django-mptt==0.18.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.3
gunicorn==23.0.0
idna==3.10
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.9
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
stripe==10.7.0
typing_extensions==4.15.0
urllib3==2.5.0
